display_groups: decode_display_group_updated proto-only at floor 213 (PR-C5)
Gate is PROTOBUF_REST_MESSAGES_3 (213), the connection floor since #632.
Collapse the PR-A dual-format wrapper: receiver flipped &mut ResponseMessage
→ &ResponseMessage via require_proto(); drop log::warn import + decode_proto_or_text
plumbing; decode_display_group_updated_proto is now the single decode path.

Add DisplayGroupUpdatedResponse builder in new src/testdata/builders/display_groups.rs.
Extract inline tests to sibling *_tests.rs files (rule 8/9): decoders_tests.rs,
stream_decoders_tests.rs, sync_tests.rs, async_tests.rs. Migrate every text-frame
fixture to proto_response(IncomingMessages::DisplayGroupUpdated, ...encode_proto()).
Add test_decode_display_group_updated_rejects_text_framing regression test.

Mark ResponseMessage::len #[allow(dead_code)] (last production caller was the
text branch deleted above. Matches PR-C2/C4 treatment of next_long / next_bool.

Update plans/legacy-text-protocol-cleanup.md to record the cleanup and drop
the now-stale display-groups callsite refs in the helper-deletion checklist.
EOF
)
