+++ plans/legacy-text-protocol-cleanup.md
@@ -76,6 +76,7 @@ Floor is now `PROTOBUF_SCAN_DATA` (210). Already-shipped deletions:
 - `decode_server_time`, `decode_server_time_millis` (gate 213 REST_MESSAGES_3) — proto-only at floor 213; receiver flipped `&mut ResponseMessage` → `&ResponseMessage` via `require_proto()`; extracted `decode_server_time_proto` / `decode_server_time_millis_proto` helpers; sync/async fixtures migrated to `proto_response(IncomingMessages::CurrentTime{,InMillis}, ...encode_proto())`; deleted text-only `server_time_test_cases` table (`invalid_timestamp` / `overflow timestamp` scenarios structurally impossible under `Option<i64>` wire); marked `ResponseMessage::next_long` `#[allow(dead_code)]` for final-cleanup deletion.
 - `decode_managed_accounts` (gate 213 REST_MESSAGES_3) — proto-only at floor 213; receiver flipped `&mut ResponseMessage` → `&ResponseMessage` via `require_proto()`; sync/async fixtures migrated from `managed_accounts().accounts(...).encode_pipe()` text frames to `proto_response(IncomingMessages::ManagedAccounts, ...encode_proto())`; deleted redundant `managed_accounts_test_cases` table (happy-path edge cases covered by existing `decode_managed_accounts_proto*` unit tests); added `_rejects_text_framing` regression test.
 - `decode_market_depth_exchanges` (gate 213 REST_MESSAGES_3) — proto-only at floor 213; receiver flipped `&mut ResponseMessage` → `&ResponseMessage` via `require_proto()`; dropped `server_version` arg (legacy text branch swapped service-data-type encoding pre-SERVICE_DATA_TYPE); added `MktDepthExchangesResponse` + `DepthMarketDataDescriptionFields` builders in `src/testdata/builders/market_data.rs`; sync/async fixtures migrated from raw `"71|2|..."` text frames to `proto_response(IncomingMessages::MktDepthExchanges, ...encode_proto())`; added `_rejects_text_framing` regression test; marked `ResponseMessage::next_bool` `#[allow(dead_code)]` for final-cleanup deletion.
+- `decode_display_group_updated` (gate 213 REST_MESSAGES_3) — proto-only at floor 213; collapsed the PR-A dual-format wrapper, receiver flipped `&mut ResponseMessage` → `&ResponseMessage` via `require_proto()`; added `DisplayGroupUpdatedResponse` builder in new `src/testdata/builders/display_groups.rs`; sync/async tests + decoder tests extracted to sibling `*_tests.rs` files (rule 8) and migrated from text frames to `proto_response(IncomingMessages::DisplayGroupUpdated, ...encode_proto())`; added `_rejects_text_framing` regression test.
 - `decode_head_timestamp`, `decode_historical_data`, `decode_historical_data_end`, `decode_historical_data_update`, `decode_historical_schedule`, `decode_historical_ticks_bid_ask`, `decode_historical_ticks_mid_point`, `decode_historical_ticks_last`, `decode_histogram_data` (all gate 208 PROTOBUF_HISTORICAL_DATA) — proto-only at floor 210; receiver flipped `&mut ResponseMessage` → `&ResponseMessage` (matches scanner/news shape); dropped `time_zone` arg from `decode_head_timestamp` / `decode_historical_data_update`, dropped `server_version` + `time_zone` from `decode_historical_data` / `decode_historical_data_end`; deleted helpers `parse_date` and `parse_bar_date` (text-only); removed `time_zone(&Client)` helpers from both `historical/sync.rs` and `historical/async.rs`; added 9 response builders (`HeadTimestampResponse`, `HistoricalDataResponse`, `HistoricalDataEndResponse`, `HistoricalDataUpdateResponse`, `HistoricalScheduleResponse`, `HistoricalTicksResponse`, `HistoricalTicksLastResponse`, `HistoricalTicksBidAskResponse`, `HistogramDataResponse`) in `src/testdata/builders/market_data.rs`; marked `ResponseMessage::next_date_time*`, `parse_ib_date_time_with_timezone`, `resolve_primitive_date_time` `#[allow(dead_code)]` for final-cleanup deletion.
 
 Decoders whose text branch is now unreachable at floor 210 and can be deleted
@@ -88,7 +89,6 @@ originating outgoing-request gate is > 210:
 
 - `decode_next_valid_id` — `RequestIds` and `StartApi` handshake (gate 213)
 - WSH event data decoders — `RequestWshEventData` (REST batch ≥ 211)
-- Display groups decoders — `QueryDisplayGroups` etc. (REST batch ≥ 211)
 
 Each follow-up PR should ground its family's response-format mapping in
 captured wire data before deleting; 213 (`PROTOBUF_REST_MESSAGES_3`) is the
@@ -123,7 +123,6 @@ Production sentinels have moved off `ResponseMessage::from(&str)` — the
 `From<Error> for RoutedItem` impl in `subscriptions/common.rs:67`. Remaining
 production callers of `ResponseMessage::from(&str)` are limited to:
 
-- `src/display_groups/common/decoders.rs:41` and `src/display_groups/common/stream_decoders.rs:51` — wrapping a parsed text payload after server-side framing; replace when display groups gets a proto decoder.
 - `src/connection/common.rs:387` — text-path branch of `parse_raw_message`; dead at floor 210 (server_version < `PROTOBUF` cannot occur).
 - `src/stubs.rs:99` — test-fixture-only (the legacy `with_responses(Vec<String>)` path).
