+++ src/display_groups/async.rs
@@ -91,102 +91,5 @@ impl Client {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::common::test_utils::helpers::assert_proto_msg_id;
-    use crate::messages::OutgoingMessages;
-    use crate::stubs::MessageBusStub;
-    use crate::subscriptions::SubscriptionItem;
-    use futures::StreamExt;
-    use std::sync::{Arc, RwLock};
-
-    #[tokio::test]
-    async fn test_subscribe_to_group_events() {
-        let message_bus = Arc::new(MessageBusStub {
-            request_messages: RwLock::new(vec![]),
-            // DisplayGroupUpdated (68), version 1, reqId 9000, contractInfo "265598@SMART"
-            response_messages: vec!["68\x001\x009000\x00265598@SMART\x00".to_string()],
-            ordered_responses: vec![],
-        });
-
-        let client = Client::stubbed(message_bus.clone(), 176);
-
-        let mut subscription = client.subscribe_to_group_events(1).await.expect("failed to subscribe");
-
-        // Verify request was sent
-        {
-            let requests = message_bus.request_messages.read().unwrap();
-            assert_eq!(requests.len(), 1);
-            assert_proto_msg_id(&requests[0], OutgoingMessages::SubscribeToGroupEvents);
-        }
-
-        // Verify response
-        let Some(Ok(SubscriptionItem::Data(update))) = subscription.next().await else {
-            panic!("expected Data");
-        };
-        assert_eq!(update.contract_info, "265598@SMART");
-    }
-
-    #[tokio::test]
-    async fn test_subscribe_to_group_events_empty_group() {
-        let message_bus = Arc::new(MessageBusStub {
-            request_messages: RwLock::new(vec![]),
-            response_messages: vec!["68\x001\x009000\x00".to_string()],
-            ordered_responses: vec![],
-        });
-
-        let client = Client::stubbed(message_bus, 176);
-
-        let mut subscription = client.subscribe_to_group_events(2).await.expect("failed to subscribe");
-
-        let Some(Ok(SubscriptionItem::Data(update))) = subscription.next().await else {
-            panic!("expected Data");
-        };
-        assert_eq!(update.contract_info, "");
-    }
-
-    #[tokio::test]
-    async fn test_update_display_group() {
-        let message_bus = Arc::new(MessageBusStub {
-            request_messages: RwLock::new(vec![]),
-            // Need a response so subscription can be created
-            response_messages: vec!["68\x001\x009000\x00265598@SMART\x00".to_string()],
-            ordered_responses: vec![],
-        });
-
-        let client = Client::stubbed(message_bus.clone(), 176);
-
-        let subscription = client.subscribe_to_group_events(1).await.expect("failed to subscribe");
-        subscription.update("265598@SMART").await.expect("update failed");
-
-        let requests = message_bus.request_messages.read().unwrap();
-        // First request is subscribe, second is update
-        assert_eq!(requests.len(), 2);
-        assert_proto_msg_id(&requests[0], OutgoingMessages::SubscribeToGroupEvents);
-        assert_proto_msg_id(&requests[1], OutgoingMessages::UpdateDisplayGroup);
-    }
-
-    #[tokio::test]
-    async fn test_subscribe_to_group_events_skips_wrong_message_type() {
-        let message_bus = Arc::new(MessageBusStub {
-            request_messages: RwLock::new(vec![]),
-            // First message is DisplayGroupList (67) - wrong type, should be skipped
-            // Second message is DisplayGroupUpdated (68) - correct type
-            response_messages: vec![
-                "67\x001\x009000\x00wrong message\x00".to_string(),
-                "68\x001\x009000\x00correct message\x00".to_string(),
-            ],
-            ordered_responses: vec![],
-        });
-
-        let client = Client::stubbed(message_bus, 176);
-
-        let mut subscription = client.subscribe_to_group_events(1).await.expect("failed to subscribe");
-
-        // Should skip the wrong message type and return the correct one
-        let Some(Ok(SubscriptionItem::Data(update))) = subscription.next().await else {
-            panic!("expected Data");
-        };
-        assert_eq!(update.contract_info, "correct message");
-    }
-}
+#[path = "async_tests.rs"]
+mod tests;
