+++ src/display_groups/common/stream_decoders.rs
@@ -43,68 +43,5 @@ impl StreamDecoder<DisplayGroupUpdate> for DisplayGroupUpdate {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn make_response(fields: &[&str]) -> ResponseMessage {
-        let raw = fields.join("\0") + "\0";
-        ResponseMessage::from(&raw)
-    }
-
-    fn test_context() -> DecoderContext {
-        DecoderContext::new(176, None)
-    }
-
-    #[test]
-    fn test_decode_display_group_update() {
-        let mut message = make_response(&["68", "1", "9000", "265598@SMART"]);
-
-        let result = DisplayGroupUpdate::decode(&test_context(), &mut message).expect("decoding failed");
-
-        assert_eq!(result.contract_info, "265598@SMART");
-    }
-
-    #[test]
-    fn test_decode_display_group_update_empty() {
-        let mut message = make_response(&["68", "1", "9000"]);
-
-        let result = DisplayGroupUpdate::decode(&test_context(), &mut message).expect("decoding failed");
-
-        assert_eq!(result.contract_info, "");
-    }
-
-    #[test]
-    fn test_decode_wrong_message_type() {
-        let mut message = make_response(&["67", "1", "9000", "data"]);
-
-        let result = DisplayGroupUpdate::decode(&test_context(), &mut message);
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_decode_error_message_surfaces_tws_error() {
-        // Error on the request_id channel surfaces as Error::Notice, not silently
-        // skipped via UnexpectedResponse (#434).
-        use crate::common::test_utils::helpers::assert_tws_error_message;
-        let mut message = make_response(&["4", "2", "9000", "10089", "Requested market data is not subscribed"]);
-        let err = DisplayGroupUpdate::decode(&test_context(), &mut message).unwrap_err();
-        assert_tws_error_message(err, 10089, "not subscribed");
-    }
-
-    #[test]
-    fn test_cancel_message() {
-        use crate::common::test_utils::helpers::assert_proto_msg_id;
-        use crate::messages::OutgoingMessages;
-
-        let message = DisplayGroupUpdate::cancel_message(176, Some(9000), None).expect("cancel message failed");
-        assert_proto_msg_id(&message, OutgoingMessages::UnsubscribeFromGroupEvents);
-    }
-
-    #[test]
-    fn test_cancel_message_no_request_id() {
-        let result = DisplayGroupUpdate::cancel_message(176, None, None);
-
-        assert!(result.is_err());
-    }
-}
+#[path = "stream_decoders_tests.rs"]
+mod tests;
